requirements.txt
beautifulsoup4==4.11.1
fastapi==0.95.1
python-dotenv==1.0.0
Requests==2.30.0
starlette==0.26.1
uvicorn==0.22.0
